+++ CHANGELOG.md
@@ -5,6 +5,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
 ## Unreleased
+-   Added: Fragmentable::beforeCreate() callback to alter fragment before the create view.
 
 ## 0.8.11 - 2024-01-16
 -   Fixed: display of duplicate options on edit.